controllers refactored and made modular to facilitate multiple async calls where required dbApi and auth files created within controllers folder.

+++ client/src/components/Provider/ProviderRemove.js
@@ -1,10 +1,12 @@
 import React, { Component } from 'react';
+import { reduxForm } from 'redux-form';
 import { connect } from 'react-redux';
 import { withStyles, Card, Typography } from '@material-ui/core';
 import BtnAction from '../UI/Buttons/btnAction'
 import DialogCustom from '../UI/Dialogs/dialogCustom';
 import CallBack from '../UI/callback'
 import HrStyled from '../UI/hrStyled'
+import FormText from '../UI/Forms/formText'
 import { selectConsoleTitle } from '../../actions'
 import providerAPI from "../../utils/provider";
 import ProviderDetailsBar from './ProviderDetailsBar'
@@ -28,17 +30,20 @@ class ProviderRemove extends Component {
         failed: false,
     }
 
-    handleRemove = () => {
-        providerAPI.delete(this.props.provider._id)
-        .then(res => {
-            console.log("res.data: ", res.data)
-            this.setState ({success: true})   // update success dialog
-        })
-        .catch(err => {
-            console.log(`OOPS! A fatal problem occurred and your request could not be completed`);
-            console.log(err);
-            this.setState({failed: true}); // update failed dialog
-        })
+    submit = (values) => {
+        console.log(values)
+        if (values.lastname === this.props.provider.lastname) {
+            providerAPI.delete(this.props.provider._id)
+            .then(res => {
+                console.log("result: ", res.data)
+                this.setState ({success: true})   // update success dialog
+            })
+            .catch(err => {
+                console.log(err)
+                console.log(err.response)
+                this.setState({failed: true}); // update failed dialog
+            })
+        } else this.props.reset("Confirmform")
     }
 
     handleCancel = () => {
@@ -49,18 +54,18 @@ class ProviderRemove extends Component {
 
   
     render () {
-        const { provider, error, loading, classes } = this.props
+        const { provider, error, loading, classes, handleSubmit } = this.props
         const { failed, success} = this.state  
 
         const texts=["Select 'Remove' to remove this provider from the list of providers held in the application. 'Cancel' to cancel.",
                      "Note, this action cannot be undone. Removed providers can be added back by re-entering all their details via the 'new provider' page."]
         
         if (error) {
-            return <div>Error! {error.message}</div>
+            return <Typography variant="subtitle1">"Care group cannot be deleted at this time. This may be temporary, so please try again later"/></Typography>
         }
 
-        if (loading || !provider._id) {
-            return <CallBack />
+        if (loading || !(provider && provider._id)) {
+            return <CallBack fallBackText="Care group cannot be deleted at this time. This may be temporary, so please try again later"/>
         }
 
 
@@ -73,10 +78,17 @@ class ProviderRemove extends Component {
                 <Typography variant="subtitle1" gutterBottom color="error">{texts[1]}</Typography>
 
                 <br /> <HrStyled /> <br />
-                            
-                <BtnAction type ="button" disabled={false} text="cancel" marginRight={true} handleAction={this.handleCancel} />
-                <BtnAction type="button" disabled={false} text="delete" warning={true} handleAction={this.handleRemove} />
 
+                <form noValidate onSubmit={handleSubmit(this.submit.bind(this))}>
+                <span style={{ marginRight: "20px"}}> 
+                    <FormText type="text" name="lastname" label="Enter provider lastname to remove" variant="outlined" width="320" helpText={false} />
+                </span>
+                <span style={{position: "relative", top: "38px"}}>
+                    <BtnAction type="submit" disabled={false} text="remove" marginRight={true} warning={true} />
+                    <BtnAction type ="button" disabled={false} text="cancel" marginRight={true} handleAction={this.handleCancel} />
+                </span>
+                </form>
+                
                 {success && 
                     <DialogCustom title="Success!" width="600px">
                         <Typography variant="subtitle1">
@@ -88,7 +100,7 @@ class ProviderRemove extends Component {
                             successfully removed from the application.
                         </Typography>
                         <br /><br />
-                        <BtnAction text="close" style={{float: "right"}} handleAction={() => this.handleCancel()} />
+                        <BtnAction text="close" style={{float: "right"}} handleAction={this.handleCancel} />
                     </DialogCustom>
                 }
 
@@ -101,10 +113,10 @@ class ProviderRemove extends Component {
                                 firstname={provider.firstname} 
                                 lastname={provider.lastname} 
                             />
-                            ' could not be deleted at this time. Please check that this is an appropriate action and try again if required. If the problem persists, contact the system administrator.
+                            ' could not be removed at this time. Please check that this is an appropriate action and try again if required. If the problem persists, contact the system administrator.
                         </Typography>
                         <br /><br />
-                        <BtnAction text="close" handleAction={() => this.handleCancel()} />
+                        <BtnAction text="close" handleAction={this.handleCancel} />
                         <br/>
                     </DialogCustom>
                 }
@@ -114,6 +126,9 @@ class ProviderRemove extends Component {
     }
 }
 
+const formData = {
+	form: 'ConfirmForm', //unique identifier for this form
+}
 
 const mapStateToProps = (state) => {
     console.log("State : ", state);
@@ -125,5 +140,6 @@ const mapStateToProps = (state) => {
 };
 
 ProviderRemove = withStyles(styles)(ProviderRemove)
+ProviderRemove = reduxForm(formData)(ProviderRemove)
 ProviderRemove = connect(mapStateToProps)(ProviderRemove)
 export default ProviderRemove;
